пред-отчетные изменение и решение лаб. раб. 8

+++ unioversity_prog/lab_008_check_sum.h
@@ -1,3 +1,3 @@
 #pragma once
 #include "lab_008_simple_value.h"
-int checkSum(int size, int* arr);
+int checkSum(int size, long int* arr);
